Move `break_assignments_in_seqs`
